Update Pokemon Wordle.js
A complete overhaul. The old code was difficult to follow and had lots of dead weight. I have refactored some of the easier to implement changes such as a rest function that works, added buttons for a reset and start-game. I have also made it so these buttons only work in certain situations.

I have made it so the dropdown list of options are now dynamically updated as and when the user starts to type instead of showing the complete list of options all at once. The dropdown items appear and disappear based on whether the name matches what the "substr()" of the user has typed.

The dropdown boxes are able to be clicked to autofill the input to save the user from needing to type the entire name out.

The list of options are made with only a few lines of code now instead of a manually typed html <datalist> which is much more maintainable if/when the options array gets updated or the use of an API is implemented, it also just gets rid of bloat.

I'm really happy with this as it stands and would only like to update the CSS at a later time.